==> notebooks/05_preparation_des_donnees.ipynb <==
# Initialiser Spark

from pyspark import SparkContext, SparkConf
conf = SparkConf().setAppName("Notebook").setMaster("local")
sc = SparkContext(conf=conf)

# Vérifier le contenu du fichier

prenoms = sc.textFile("../data/prenoms.txt")
prenoms.takeSample(False, 10)

# Extraire et sauvegarder un échantillon
1. Expliquer ce que fait la ligne suivante

prenoms.filter(lambda l: l.startswith("sexe") == False)\
    .filter(lambda l: "XX" not in l)\
    .sample(False, .001)\
    .saveAsTextFile("../data/prenoms_sample.txt")

Cette ligne filtre les lignes qui commencent par "sexe" (elle enlève l'en tête), elle filtre aussi les lignes contenant "XX" (lignes avec des données manquantes) puis elle prends un échantillon aléatoire contenant 0.1% des lignes et elle le sauvegarde dans le dossier prenoms_sample.txt.

# Vérifier l'échantillon

echantillon = sc.textFile("../data/prenoms_sample.txt")
print(echantillon.count())
print(echantillon.filter(lambda l: l.startswith("sexe")).count())
print(echantillon.filter(lambda l: "XX" in l).count())
