==> notebooks/10_introduction_aux_rdd.ipynb <==
# Initialisation de Spark

from pyspark import SparkContext, SparkConf
conf = SparkConf().setAppName("Notebook").setMaster("local")
sc = SparkContext(conf=conf)

# Une fonction pour simplifier l'accès aux données

def extract_field(s: str, field_number: int, sep: str = ';') -> str:
    fields = s.split(sep)
    return fields[field_number] if (field_number < len(fields)) else ""

print(extract_field("2;CASSIOPEE;2009;33;3", 0))
print(extract_field("2;CASSIOPEE;2009;33;3", 1))
print(extract_field("2;CASSIOPEE;2009;33;3", 2))
print(extract_field("2;CASSIOPEE;2009;33;3", 3))
print(extract_field("2;CASSIOPEE;2009;33;3", 4))
print(extract_field("2;CASSIOPEE;2009;33;3", 5))

# Charger les données
1. Créer le RDD `lignes` à partir du répertoire `prenoms_sample.txt`

lignes = sc.textFile("../data/prenoms_sample.txt")
print(lignes.take(10))

# Transformer les lignes en prénoms
1. En appliquant la méthode `map`, créer le RDD `prenoms` à partir de `lignes`

prenoms = lignes.map(lambda l: (
    int(extract_field(l, 0)),
    extract_field(l, 1),
    int(extract_field(l, 2)),
    extract_field(l, 3),
    int(extract_field(l, 4))
))
for n in prenoms.take(10):
    print(n)

# Interroger les données
La documentation des méthodes d'un RDD est disponible ([RDD](https://spark.apache.org/docs/latest/api/python/reference/api/pyspark.RDD.html)).

1. Rappeler ce que sont les *transformations* et les *actions*
1. Donner, pour chaque prénom, son nombre d'occurences (`map` et `reduceByKey`)

Transformation : transforme le RDD courant et retourne un nouveau RDD.  
Action : exécute le calcul d'une valeur sur un RDD.

from operator import add
prenoms.map(lambda t: (t[1], 1)).reduceByKey(add).collect()

1. Donner le nombre total de naissances avec un prénom féminin (`filter`, `map`, `reduce` ou `sum`)

prenoms.filter(lambda t : t[0] == 2).map(lambda t: t[4]).sum()

1. Donner l'effectif maximal et minimal par prénom (`map`, `aggregateByKey`)

prenoms.map(lambda t: (t[1], t[4])).aggregateByKey(
    (float('inf'), float('-inf')),
    lambda acc, v: (min(acc[0], v), max(acc[1], v)),
    lambda acc1, acc2: (min(acc1[0], acc2[0]), max(acc1[1], acc2[1]))
).collect()

1. Sur le modèle des prénoms, charger les données des départements
1. Donner, pour chaque nom de département, le prénom le plus fréquent depuis l'année 2000

lignes = sc.textFile("../data/dpts.txt")
header = lignes.first()
lignes_sans_entete = lignes.filter(lambda l: l != header)

dpts = lignes_sans_entete.map(lambda l: (
    extract_field(l, 0, ","),
    int(extract_field(l, 1, ",")),
    extract_field(l, 2, ","),
    int(extract_field(l, 3, ",")),
    extract_field(l, 4, ","),
    extract_field(l, 5, ","),
    extract_field(l, 6, ",")
))

for n in dpts.take(10):
    print(n)

prenoms.filter(lambda t: int(t[2]) >= 2000) \
        .map(lambda t: ((t[3], t[1]), t[4])).reduceByKey(add) \
        .map(lambda t: (t[0][0], (t[0][1], t[1]))) \
        .reduceByKey(lambda a, b: a if a[1] >= b[1] else b) \
        .map(lambda t: (t[0], t[1][0])) \
        .join(dpts.map(lambda t: (t[0], t[6]))) \
        .map(lambda t: (t[1][1], t[1][0])) \
        .collect()
