==> notebooks/25_comparaison_des_formats.ipynb <==
# Initialisation de Spark

from pyspark.sql import SparkSession
spark = SparkSession \
    .builder \
    .appName("Exemples avec les formats de stockage") \
    .getOrCreate()
sc = spark.sparkContext

# Chargement des données

from pyspark.sql import Row

# Charge un fichier texte et convertit les lignes en "Row".
lines = sc.textFile("../data/prenoms.txt")
prenoms_as_rdd = lines.map(lambda l: l.split(";"))\
    .filter(lambda l: "sexe" not in l and "XX" not in l)\
    .map(lambda p: Row(\
        sexe=int(p[0]),\
        prenom=p[1],\
        annee=int(p[2]),\
        dep=p[3],\
        effectif=int(p[4])))

# Infère le schéma et enregistre le DataFrame comme une table.
prenoms = spark.createDataFrame(prenoms_as_rdd)
prenoms.cache()

# Sauvegarder dans les différents formats

import os

def get_size(start_path):
    total_size = 0
    for dirpath, dirnames, filenames in os.walk(start_path):
        for f in filenames:
            fp = os.path.join(dirpath, f)
            if not os.path.islink(fp): # skip if it is symbolic link
                total_size += os.path.getsize(fp)
    return total_size

formats = {\
    "csv": ["uncompressed", "bzip2", "deflate", "gzip"],\
    "json": ["uncompressed", "bzip2", "deflate", "gzip"],\
    "parquet": ["uncompressed", "gzip", "snappy"],\
    "orc": ["uncompressed", "snappy", "zlib"]\
}

size_by_formats_and_codecs = dict()
for format in iter(formats):
    size_by_formats_and_codecs[format] = dict()
    for codec in formats[format]:
        prenoms.write\
            .mode("overwrite")\
                .option("compression", codec)\
                    .format(format)\
                        .save("../data/prenoms")
        size = get_size("../data/prenoms")
        print(f'{format}, {codec}, {size}')
        size_by_formats_and_codecs[format][codec] = size
print(size_by_formats_and_codecs)

# Affichage des résultats

## Taille (en Mo)

all_codecs = ["uncompressed", "bzip2", "deflate", "gzip", "snappy", "zlib"]
print("Format | uncomp.  | bzip2    | def.     | gzip     | snappy   | zlib     |")
for format in iter(formats):
    print(f'{format:7}|', end='')
    for codec in all_codecs:
        size = size_by_formats_and_codecs[format].get(codec, 0) / (1024 * 1024)
        print(f' {size:8.2f} |', end='')
    print()
