==> notebooks/30_spark_sql.ipynb <==
# Initialisation de Spark

from pyspark.sql import SparkSession
spark = SparkSession \
    .builder \
    .appName("Exemples avec les formats de stockage") \
    .getOrCreate()
sc = spark.sparkContext

# Chargement des données
1. Créer un dataframe pour les prénoms

prenoms = spark.read.load("../data/prenomsParDeptsEtAnnees.parquet")
prenoms.createOrReplaceTempView("prenoms")
prenoms.show()

1. Faire de même pour les départements

depts = spark.read.load("../data/depts.parquet")
depts.createOrReplaceTempView("depts")
depts.show()

# Interroger les données

* La documentation sur Spark SQL est disponible [ici](https://spark.apache.org/docs/latest/sql-programming-guide.html) (de même pour [DataFrame](https://spark.apache.org/docs/latest/api/python/reference/pyspark.sql/api/pyspark.sql.DataFrame.html))
* Pour chaque question, vous donnerez la réponse avec le DSL et en SQL

1. Donner, pour chaque prénom, son nombre d'occurences

## DSL

prenoms.groupBy("prenom").count().show()

## SQL

spark.sql('SELECT prenom, COUNT(*) FROM prenoms GROUP BY prenom').show()

1. Donner le nombre total de naissances avec un prénom féminin

## DSL

prenoms.filter(prenoms.sexe == 2).groupBy("sexe").count().show()

## SQL

spark.sql("SELECT COUNT(*) FROM prenoms WHERE sexe = 2").show()

1. Donner l'effectif maximal et minimal par prénom

## DSL

from pyspark.sql.functions import min, max

prenoms.groupBy("prenom").agg(min("effectif"), max("effectif")).show()

## SQL

spark.sql("SELECT prenom, MIN(effectif), MAX(effectif) FROM prenoms GROUP BY prenom").show()

1. Donner, pour chaque nom de département, le nombre moyen de prénoms depuis l'année 2000

## DSL

from pyspark.sql.functions import mean

prenoms.filter(prenoms.annee >= 2000).join(depts, prenoms.dep == depts.dep).groupBy("libelle").agg(mean("effectif")).show()

## SQL

spark.sql("SELECT libelle, AVG(effectif) FROM prenoms, depts WHERE prenoms.dep = depts.dep AND annee >= 2000 GROUP BY depts.libelle").show()
